+++ BoostAsioServer/ServerCore/RecvBuffer.h
@@ -5,23 +5,23 @@ class RecvBuffer
 	enum { BUFFER_COUNT = 10 };
 
 public:
-	RecvBuffer(int32 bufferSize);
+	RecvBuffer(uint64 bufferSize);
 	~RecvBuffer();
 
 	void			Clean();
-	bool			OnRead(int32 numOfBytes);
-	bool			OnWrite(int32 numOfBytes);
+	bool			OnRead(uint64 numOfBytes);
+	bool			OnWrite(uint64 numOfBytes);
 
 	BYTE* ReadPos() { return &_buffer[_readPos]; }
 	BYTE* WritePos() { return &_buffer[_writePos]; }
-	int32			DataSize() { return _writePos - _readPos; }
-	int32			FreeSize() { return _capacity - _writePos; }
+	uint64			DataSize() { return _writePos - _readPos; }
+	uint64			FreeSize() { return _capacity - _writePos; }
 
 private:
-	int32				_capacity = 0;
-	int32				_bufferSize = 0;
-	int32				_readPos = 0;
-	int32				_writePos = 0;
+	uint64				_capacity = 0;
+	uint64				_bufferSize = 0;
+	uint64				_readPos = 0;
+	uint64				_writePos = 0;
 	std::vector<BYTE>	_buffer;
 };
